api/pyproject.toml
[tool.poetry]
name = "urbantech-test-task"
version = "0.1.0"
description = ""
authors = ["Максим Жданов <exodus.yarwood@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
sqlalchemy = "^2.0.32"
pydantic = "^2.8.2"
pydantic-settings = "^2.4.0"
alembic = "^1.13.2"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"
fastapi = {extras = ["standard"], version = "^0.112.0"}
python-multipart = "^0.0.9"
pillow = "^10.4.0"
redis = {extras = ["hiredis"], version = "^5.0.8"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
